-Move cart navigation link to navigation bar -Add purchased array in Content model -add subtotal function in content model

+++ Me-Moji2/Model/Model.swift
@@ -12,5 +12,5 @@ struct CustomizeItem: Identifiable, Decodable {
     var id: Int
     var name: String
     var image: String
-    var price: Float
+    var price: Int
 }
